refactor: 모델 패키지의 내의 클래스들을 적절한 패키지로 이동해라
