+++ src/main/java/lotto/model/lotto/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
