+++ src/test/java/lotto/model/LottoTest.java
@@ -1,6 +1,6 @@
 package lotto.model;
 
-import lotto.model.Lotto;
+import lotto.model.lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
