+++ src/test/java/lotto/model/LottoValidatorTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.model.ErrorMessages.Lotto;
-import lotto.model.LottoValidator;
+import lotto.model.lotto.LottoValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
