+++ src/test/java/lotto/model/RankTest.java
@@ -2,7 +2,7 @@ package lotto.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import lotto.model.Rank;
+import lotto.model.lotto.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
