+++ src/test/java/lotto/model/WinningLottoValidatorTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.ArrayList;
 import java.util.List;
 import lotto.model.ErrorMessages.Lotto;
-import lotto.model.WinningLottoValidator;
+import lotto.model.lotto.WinningLottoValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
